backend/requirements.txt
django==4.2.16
django-redis==5.4.0
djangorestframework==3.15.2
django-environ==0.9.0

uvicorn==0.34.3


channels==4.1.0
channels-redis==4.2.0


whitenoise==6.9.0
psycopg2==2.9.4


